backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
openai==1.10.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
markdown==3.5.2
weasyprint==60.2
aiosqlite==0.19.0
